requirements.txt
aws-utils==0.1
awscli==1.18.223
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
docker==4.4.1
docutils==0.15.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
pprintpp==0.4.0
pyasn1==0.4.8
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.25.1
rsa==4.5
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.3
websocket-client==0.57.0
Werkzeug==1.0.1
